pyproject.toml
# Metadata about the Python package https://python-poetry.org/docs/pyproject/
[tool.poetry]
name = "django-graphene-filters"
version = "0.1.2"
description = "Advanced auto related filters for Graphene Django"
authors = ["Rio Weber <riodweber@gmail.com>"]
maintainers = ["Rio Weber <riodweber@gmail.com>"]
license = "LICENSE"
readme = "README.md"
homepage = "https://github.com/riodw/django-graphene-filters"
repository = "https://github.com/riodw/django-graphene-filters"
keywords = ["django", "graphene", "filters", "related"]
# Classifiers helps to find project by categorizing it
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Programming Language :: Python :: 3.14",
    "Operating System :: OS Independent",
    "Framework :: Django :: 5.2",
    "Framework :: Django :: 6.0",
]

[tool.poetry.urls]
Homepage = "https://pypi.org/project/django-graphene-filters"
repository = "https://pypi.org/project/django-graphene-filters"
issues = "https://github.com/riodw/django-graphene-filters/issues"
# DOWNLOAD = "https://github.com/riodw/django-graphene-filters"

# Dependencies needed to run the package
[tool.poetry.dependencies]
python = ">=3.10,<4.0"
Django = ">=5.2"
graphene = ">=3,<4"
graphene-django = ">=3.0.0"
django-filter = "^25.2"
psycopg2-binary = "^2.9.3"
stringcase = "^1.2.0"
anytree = "^2.13.0"
wrapt = "^2.0.1"



# How the package will be built

[tool.poetry.group.dev.dependencies]
black = "^26.1.0"
coverage = "^7.13.2"
pre-commit = "^4.5.1"
pytest = "^9.0.2"
pytest-django = "^4.5.2"
django-seed = "^0.3.1"
python-dotenv = "^1.0.0"
factory-boy = "^3.3.0"
ruff = "^0.15.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110
skip-magic-trailing-comma = false
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | env
  | venv
  | \.pytest_cache
  | \.coverage
  | htmlcov
)/
'''

[tool.ruff]
line-length = 110
target-version = "py310"
exclude = [
    ".git",
    "__pycache__",
    "docs/source/conf.py",
    "old",
    "build",
    "dist",
    "venv",
    ".venv",
    "env",
    ".pytest_cache",
    ".coverage",
    "htmlcov",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "ANN", # flake8-annotations
    "A",   # flake8-builtins
    "N",   # pep8-naming
    "C4",  # flake8-comprehensions
    "D",   # pydocstyle
    "ERA", # eradicate
    "Q",   # flake8-quotes
    "SIM", # flake8-simplify
    "UP",  # pyupgrade
]
ignore = [
    "ANN002", # Missing type annotation for *args
    "ANN003", # Missing type annotation for **kwargs
    "ANN401", # Any type annotation
    "D107",   # Missing docstring in __init__
    "D417",   # Missing argument description in docstring
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["D", "ANN", "ERA001", "F841", "SIM117", "SIM118", "E501", "N802", "N806"]
"examples/**/*.py" = ["D", "ANN", "ERA001", "F841", "SIM117", "SIM118", "E501", "UP031"]
"django_graphene_filters/connection_field.py" = ["A002"]
"django_graphene_filters/filters.py" = ["A003", "D106"]
"**/migrations/*.py" = ["D100", "D101", "D104"]
"**/views.py" = ["D101"]
"**/urls.py" = ["D100"]
"**/admin.py" = ["D101"]

[tool.ruff.lint.isort]
known-first-party = ["django_graphene_filters"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"
multiline-quotes = "double"

# Semantic release settings
[tool.semantic_release]
version_variable = [
    "django_graphene_filters/__init__.py:__version__",
    "pyproject.toml:version",
]
branch = "master"
# Whether to upload to PyPI
upload_to_pypi = true
# Whether to upload to GitHub Releases
upload_to_release = true
# Build command before release
build_command = "pip install poetry && poetry build"
